Follow-up boilerplate changes * Remove redundant make targets and files * Fix golangci-lint errors * Move config.go to standard location * Update .gitignore with bp directories/files * Remove default target in Makefile
